functionality to generate ip addresses
